+++ ng/nggraphics.cpp
@@ -3,14 +3,91 @@
 
 #include "nggraphics.h"
 
-ng::Color::Color () {
-	this->r = 0;
-	this->g = 0;
-	this->b = 0;
-	this->a = 255;
+// Switch between window and graphics coordinates.
+// Window (0,0) is top-left corner, +x points right, and +y points down.
+// Graphics (0,0) is center, +x points right, and +y points up.
+// These are used internally by the graphics draw functions.
+double ng::window_x (double x, double rx) {
+	return x+rx;
 }
 
-ng::Color::~Color () {}
+double ng::window_y (double y, double ry) {
+	return -(y-ry);
+}
+
+double ng::graphics_x (double x, double rx) {
+	// window_x = graphics_x + rx
+	// graphics_x = window_x - rx
+	return x-rx;
+}
+
+double ng::graphics_y (double y, double ry) {
+	// window_y = -(graphics_y - ry)
+	// graphics_y = (-window_y) + ry
+	return (-y)+ry
+}
+
+Rect2 ng::window_rect (const Box2& a, double rx, double ry) {
+	Box2 b(ng::window_x(a.x,rx), ng::window_y(a.y,ry), a.rx, a.ry);
+	Rect2 c(b);
+	return c;
+}
+
+SDL_Rect ng::sdl_rect (const Rect2& a) {
+	SDL_Rect b;
+	b.x = static_cast<int>(c.x);
+	b.y = static_cast<int>(c.y);
+	b.w = static_cast<int>(c.w);
+	b.h = static_cast<int>(c.h);
+	return b;
+}
+
+SDL_RendererFlip ng::sdl_flip (int flip) {
+	SDL_RendererFlip flip_sdl;
+	switch (flip) {
+		case ng::FlipX:
+			flip_sdl = SDL_FLIP_HORIZONTAL;
+			break;
+		case ng::FlipY:
+			flip_sdl = SDL_FLIP_VERTICAL;
+			break;
+		case ng::FlipXY:
+			flip_sdl = static_cast<SDL_RendererFlip>(
+				static_cast<int>(SDL_FLIP_HORIZONTAL) |
+				static_cast<int>(SDL_FLIP_VERTICAL));
+		case ng::None: default:
+			flip_sdl = SDL_FLIP_NONE;
+	}
+	return flip_sdl;
+}
+
+ng::Color::Color () :
+	r(0),
+	g(0),
+	b(0),
+	a(255)
+{}
+
+ng::Color::Color (int r, int g, int b) :
+	r(r),
+	g(g),
+	b(b),
+	a(255)
+{}
+
+ng::Color::Color (int r, int g, int b, int a) :
+	r(r),
+	g(g),
+	b(b),
+	a(a)
+{}
+
+ng::Color::Color (const Color& color) :
+	r(color.r),
+	g(color.g),
+	b(color.b),
+	a(color.a)
+{}
 
 void ng::Color::set (int r, int g, int b) {
 	this->r = r;
@@ -26,13 +103,13 @@ void ng::Color::set (int r, int g, int b, int a) {
 	this->a = a;
 }
 
-ng::Image::Image () {
-	this->texture = NULL;
-	this->w = 0.0;
-	this->h = 0.0;
-	this->color.set(255, 255, 255);
-	this->flip = ng::None;
-}
+ng::Image::Image () :
+	texture(NULL),
+	w(0.0),
+	h(0.0),
+	color(255, 255, 255),
+	flip(ng::None)
+{}
 
 ng::Image::~Image () {
 	if (this->texture != NULL) {
@@ -41,7 +118,7 @@ ng::Image::~Image () {
 	}
 }
 
-void ng::Image::load (Graphics* const g, const char* file, const Color* key) {
+void ng::Image::load (Graphics* const graphics, const char* file, const Color& key) {
 	SDL_Surface* surface = NULL;
 	surface = SDL_LoadBMP(file);
 	if (surface == NULL) {
@@ -49,12 +126,12 @@ void ng::Image::load (Graphics* const g, const char* file, const Color* key) {
 	}
 	
 	if (SDL_SetColorKey(surface, SDL_TRUE, SDL_MapRGB(surface->format, 
-	key->r, key->g, key->b)) != 0) {
+	key.r, key.g, key.b)) != 0) {
 		SDL_FreeSurface(surface);
 		throw std::runtime_error(SDL_GetError());
 	}
 	SDL_Texture* texture = NULL;
-	texture = SDL_CreateTextureFromSurface(g->renderer, surface);
+	texture = SDL_CreateTextureFromSurface(graphics->renderer, surface);
 	if (texture == NULL) {
 		SDL_FreeSurface(surface);
 		throw std::runtime_error(SDL_GetError());
@@ -65,49 +142,49 @@ void ng::Image::load (Graphics* const g, const char* file, const Color* key) {
 	this->h = static_cast<double>(surface->h);
 }
 
-void ng::Image::set_color (const Color* color) {
-	if (this->color.r == color->r && this->color.g == color->g && this->color.b == color->b) {
+void ng::Image::set_color (const Color& color) {
+	if (this->color.r == color.r && this->color.g == color.g && this->color.b == color.b) {
 		return;
 	}
 	// This is a different color.
-	if (SDL_SetTextureColorMod(this->texture, color->r, color->g, color->b) != 0) {
+	if (SDL_SetTextureColorMod(this->texture, color.r, color.g, color.b) != 0) {
 		throw std::runtime_error(SDL_GetError());
 	}
-	this->color.r = color->r;
-	this->color.g = color->g;
-	this->color.b = color->b;
+	this->color.r = color.r;
+	this->color.g = color.g;
+	this->color.b = color.b;
 }
 
-void ng::Image::set_alpha (const Color* color) {
-	if (this->color.a == color->a) {
+void ng::Image::set_alpha (const Color& color) {
+	if (this->color.a == color.a) {
 		return;
 	}
 	// This is a different alpha.
-	if (SDL_SetTextureAlphaMod(this->texture, color->a) != 0) {
+	if (SDL_SetTextureAlphaMod(this->texture, color.a) != 0) {
 		throw std::runtime_error(SDL_GetError());
 	}
-	this->color.a = color->a;
+	this->color.a = color.a;
 }
 
+/*
 void ng::Image::set_flip (int flip) {
 	this->flip = flip;
 }
+*/
 
-ng::Graphics::Graphics () {
-	this->window = NULL;
-	this->renderer = NULL;
-	this->w = 0.0;
-	this->h = 0.0;
-	this->color.set(0, 0, 0);
-}
+ng::Graphics::Graphics () :
+	window(NULL),
+	renderer(NULL),
+	rx(0.0),
+	ry(0.0),
+	color(0, 0, 0)
+{}
 
-ng::Graphics::~Graphics () {}
-
-void ng::Graphics::open (const char* title, double w, double h) {
-	this->w = w;
-	this->h = h;
+void ng::Graphics::open (const char* title, double rx, double ry) {
+	this->rx = rx;
+	this->ry = ry;
 	this->window = SDL_CreateWindow(title, SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED,
-		static_cast<int>(w), static_cast<int>(h), SDL_WINDOW_RESIZABLE);
+		static_cast<int>(rx*2.0), static_cast<int>(ry*2.0), SDL_WINDOW_RESIZABLE);
 	if (this->window == NULL) {
 		throw std::runtime_error(SDL_GetError());
 	}
@@ -128,37 +205,37 @@ void ng::Graphics::close () {
 	}
 }
 
-void ng::Graphics::set_color (const Color* color) {
-	if (this->color.r == color->r && this->color.g == color->g && this->color.b == color->b) {
+void ng::Graphics::set_color (const Color& color) {
+	if (this->color.r == color.r && this->color.g == color.g && this->color.b == color.b) {
 		return;
 	}
 	// This is a different color.
 	if (SDL_SetRenderDrawColor(this->renderer,
-	color->r, color->g, color->b, this->color.a) != 0) {
+	color.r, color.g, color.b, this->color.a) != 0) {
 		throw std::runtime_error(SDL_GetError());
 	}
-	this->color.r = color->r;
-	this->color.g = color->g;
-	this->color.b = color->b;
+	this->color.r = color.r;
+	this->color.g = color.g;
+	this->color.b = color.b;
 }
 
-void ng::Graphics::set_alpha (const Color* color) {
-	if (this->color.a == color->a) {
+void ng::Graphics::set_alpha (const Color& color) {
+	if (this->color.a == color.a) {
 		return;
 	}
 	// This is a different alpha.
 	SDL_BlendMode blendmode;
-	if (color->a == 255) {
+	if (color.a == 255) {
 		blendmode = SDL_BLENDMODE_NONE;
 	} else {
 		blendmode = SDL_BLENDMODE_BLEND;
 	}
 	if (SDL_SetRenderDrawBlendMode(this->renderer, blendmode) != 0 ||
 	SDL_SetRenderDrawColor(this->renderer,
-	this->color.r, this->color.g, this->color.b, color->a) != 0) {
+	this->color.r, this->color.g, this->color.b, color.a) != 0) {
 		throw std::runtime_error(SDL_GetError());
 	}
-	this->color.a = color->a;
+	this->color.a = color.a;
 }
 
 void ng::Graphics::clear () {
@@ -173,66 +250,41 @@ void ng::Graphics::draw () {
 
 // Draw whole image to whole window.
 void ng::Graphics::draw_image (Image* const image) {
-	Rect2 src(image->w * 0.5, image->h * 0.5, image->w, image->h);
-	Rect2 dest(this->w * 0.5, this->h * 0.5, this->w, this->h);
-	this->draw_image(image, &src, &dest);
+	Rect2 src(0.0, 0.0, image->w, image->h);
+	Box2 dest(0.0, 0.0, this->rx, this->ry);
+	this->draw_image(image, src, dest);
 }
 
 // Draw whole image to part of window.
-void ng::Graphics::draw_image (Image* const image, const Rect2* dest) {
-	Rect2 src(image->w * 0.5, image->h * 0.5, image->w, image->h);
-	this->draw_image(image, &src, dest);
+void ng::Graphics::draw_image (Image* const image, const Box2& dest) {
+	Rect2 src(0.0, 0.0, image->w, image->h);
+	this->draw_image(image, src, dest);
 }
 
 // Draw part of image to part of window.
-void ng::Graphics::draw_image (Image* const image, const Rect2* src, const Rect2* dest) {
-	SDL_Rect src_sdl, dest_sdl;
-	src_sdl.x = static_cast<int>(src->p.x - (src->w * 0.5));
-	src_sdl.y = static_cast<int>(src->p.y - (src->h * 0.5));
-	src_sdl.w = static_cast<int>(src->w);
-	src_sdl.h = static_cast<int>(src->h);
-	dest_sdl.x = static_cast<int>(dest->p.x - (dest->w * 0.5));
-	dest_sdl.y = static_cast<int>(dest->p.y - (dest->h * 0.5));
-	dest_sdl.w = static_cast<int>(dest->w);
-	dest_sdl.h = static_cast<int>(dest->h);
+void ng::Graphics::draw_image (Image* const image, const Rect2& src, const Box2& dest) {
+	SDL_Rect src_sdl = ng::sdl_rect(src);
+	SDL_Rect dest_sdl = ng::sdl_rect(ng::window_rect(dest));
 	
-	if (image->flip == ng::None && src->a == 0.0 && dest->a == 0.0) {
-		if (SDL_RenderCopy(this->renderer, image->texture, &src_sdl, &dest_sdl) != 0) {
-			throw std::runtime_error(SDL_GetError());
-		}
-	} else {
-		SDL_RendererFlip flip;
-		switch (image->flip) {
-			case ng::None: {
-				flip = SDL_FLIP_NONE;
-				break;
-			} case ng::FlipX: {
-				flip = SDL_FLIP_HORIZONTAL;
-				break;
-			} case ng::FlipY: {
-				flip = SDL_FLIP_VERTICAL;
-				break;
-			} case ng::FlipXY: {
-				flip = static_cast<SDL_RendererFlip>(
-					static_cast<int>(SDL_FLIP_HORIZONTAL) |
-					static_cast<int>(SDL_FLIP_VERTICAL));
-			} default: {
-				throw std::logic_error("unsupported flip mode");
-			}
-		}
-		if (SDL_RenderCopyEx(this->renderer, image->texture, &src_sdl, &dest_sdl,
-		ng::degrees(src->a + dest->a), NULL, flip) != 0) {
-			throw std::runtime_error(SDL_GetError());
-		}
+	if (SDL_RenderCopy(this->renderer, image->texture, &src_sdl, &dest_sdl) != 0) {
+		throw std::runtime_error(SDL_GetError());
 	}
 }
 
-void ng::Graphics::draw_rect (const Rect2* dest, int draw) {
-	SDL_Rect dest_sdl;
-	dest_sdl.x = static_cast<int>(dest->p.x - (dest->w * 0.5));
-	dest_sdl.y = static_cast<int>(dest->p.y - (dest->h * 0.5));
-	dest_sdl.w = static_cast<int>(dest->w);
-	dest_sdl.h = static_cast<int>(dest->h);
+void ng::Graphics::draw_image (Image* const image, const Rect2& src, const Box2& dest,
+double angle, int flip) {
+	SDL_Rect src_sdl = ng::sdl_rect(src);
+	SDL_Rect dest_sdl = ng::sdl_rect(ng::window_rect(dest));
+	
+	if (SDL_RenderCopyEx(this->renderer, image->texture, &src_sdl, &dest_sdl,
+	ng::degrees(angle), NULL, ng::sdl_flip(flip)) != 0) {
+		throw std::runtime_error(SDL_GetError());
+	}
+}
+
+// Draw shape.
+void ng::Graphics::draw_box (const Box2& dest, int draw) {
+	SDL_Rect dest_sdl = ng::sdl_rect(ng::window_rect(dest));
 	
 	switch (draw) {
 		case ng::DrawFill: {
@@ -240,33 +292,30 @@ void ng::Graphics::draw_rect (const Rect2* dest, int draw) {
 				throw std::runtime_error(SDL_GetError());
 			}
 			break;
-		} case ng::DrawFrame: {
+		} case ng::DrawFrame: default: {
 			if (SDL_RenderDrawRect(this->renderer, &dest_sdl) != 0) {
 				throw std::runtime_error(SDL_GetError());
 			}
-			break;
-		} default: {
-			throw std::logic_error("unsupported draw mode");
 		}
 	}
 }
 
-void ng::Graphics::draw_line (const Line2* line) {
+void ng::Graphics::draw_line (const Vec2& p1, const Vec2& p2) {
 	int x1, y1, x2, y2;
-	x1 = static_cast<int>(line->p.x);
-	y1 = static_cast<int>(line->p.y);
-	x2 = static_cast<int>(line->p.x + line->v.x);
-	y2 = static_cast<int>(line->p.y + line->v.y);
+	x1 = static_cast<int>(ng::window_x(p1.x, this->rx));
+	y1 = static_cast<int>(ng::window_y(p1.y, this->ry));
+	x2 = static_cast<int>(ng::window_x(p2.x, this->rx));
+	y2 = static_cast<int>(ng::window_y(p2.y, this->ry));
 
 	if (SDL_RenderDrawLine(this->renderer, x1, y1, x2, y2) != 0) {
 		throw std::runtime_error(SDL_GetError());
 	}
 }
 
-void ng::Graphics::draw_point (const Vec2* vec) {
+void ng::Graphics::draw_point (const Vec2& p) {
 	int x, y;
-	x = static_cast<int>(vec->x);
-	y = static_cast<int>(vec->y);
+	x = static_cast<int>(ng::window_x(p.x, this->rx));
+	y = static_cast<int>(ng::window_y(p.y, this->ry));
 
 	if (SDL_RenderDrawPoint(this->renderer, x, y) != 0) {
 		throw std::runtime_error(SDL_GetError());